modify api Put and delete

+++ EndpointAPI/PokemonEndpoint.cs
@@ -22,19 +22,29 @@ public static class PokemonEndpoint
             return result is null ? Results.BadRequest(new { message = "Pokemon create unsuccessful." }) : Results.Ok(result);
         });
 
-        app.MapPut("/pokemonUpdate", async (IPokemonService pokemonService, [FromBody] PokemonDTO pokemon) =>
+        app.MapPut("/pokemonUpdate", async (IPokemonService pokemonService, string id, [FromBody] PokemonDTO pokemon) =>
         {
-            var result = await pokemonService.UpdatePokemonAsync(pokemon);
+            if (string.IsNullOrEmpty(id) || string.IsNullOrEmpty(pokemon?.Name))
+            {
+                return Results.BadRequest(new { message = "Pokemon update unsuccessful." });
+            }
+
+            var result = await pokemonService.UpdatePokemonAsync(id, pokemon);
 
             return result is null ? Results.BadRequest(new { message = "Pokemon update unsuccessful." }) : Results.Ok(result);
         });
 
         app.MapDelete("/pokemonDelete", async (IPokemonService pokemonservice, string id) =>
         {
+            if (string.IsNullOrEmpty(id))
+            {
+                return Results.BadRequest(new { message = "Pokemon delete unsuccessful." });
+            }
 
             var result = await pokemonservice.DeletePokemonAsync(id);
 
-            return result is false ? Results.NoContent() : Results.Ok(new { message = "Pokemon successfully deleted." }); ;
+            return result is false ? Results.BadRequest(new { message = "Pokemon delete unsuccessful." })
+                                     : Results.Ok(new { message = "Pokemon successfully deleted." }); ;
 
         });
     }
